refactor: merge sub-modules of io._dict
