--- src/expertsystem/io/_dict/__init__.py
@@ -1,10 +1,0 @@
-"""Serialization from and to a `dict`."""
-
-
-from . import build, dump, validate
-
-__all__ = [
-    "build",
-    "dump",
-    "validate",
-]
